requirements.txt
# NOTE: the following is needed only for creating a list of English words.
requests==2.31.0
pyenchant==3.2.2
